+++ Java/src/main/java/com/gildedrose/item_helpers/ItemHandler.java
@@ -1,15 +1,14 @@
 package com.gildedrose.item_helpers;
 
-import com.gildedrose.Item;
+import com.gildedrose.main.Item;
 
-import static com.gildedrose.item_helpers.ItemName.LEGENDARY;
+import static com.gildedrose.items.LegendaryItem.NOT_LEGENDARY_ITEM_ERROR_MESSAGE;
+import static com.gildedrose.items.LegendaryItem.isNotLegendary;
 import static java.lang.Math.max;
 
 public class ItemHandler {
 
-    private static final int LEGENDARY_ITEM_QUALITY = 80;
-    protected static final String NEGATIVE_QUALITY_ERROR_MESSAGE = "Quality cannot be negative! Current value: ";
-    protected static final String NOT_LEGENDARY_ITEM_ERROR_MESSAGE = "Item is legendary, quality must be always 80! Current value: ";
+    protected static final String QUALITY_ERROR_MESSAGE = "Quality cannot be negative! Current value: ";
 
     private final Item item;
 
@@ -19,7 +18,7 @@ public class ItemHandler {
 
     static void validate(Item item) {
         if (qualityIsNegative(item)) {
-            throw new IllegalArgumentException(NEGATIVE_QUALITY_ERROR_MESSAGE + item.quality);
+            throw new IllegalArgumentException(QUALITY_ERROR_MESSAGE + item.quality);
         }
         if (isNotLegendary(item)) {
             throw new IllegalArgumentException(NOT_LEGENDARY_ITEM_ERROR_MESSAGE + item.quality);
@@ -30,8 +29,12 @@ public class ItemHandler {
         return item.quality < 0;
     }
 
-    private static boolean isNotLegendary(Item item) {
-        return item.name.equals(LEGENDARY.toString()) && item.quality != 80;
+    public boolean qualityIsHigherThanZero() {
+        return item.quality > 0;
+    }
+
+    public void makeQualityZero() {
+        item.quality = 0;
     }
 
     public void decrementSellInDate() {
@@ -42,32 +45,6 @@ public class ItemHandler {
         return item.sellIn >= 0;
     }
 
-    public void incrementQualityByTwo() {
-        item.quality = max(item.quality + 2, 0);
-    }
-
-    public void incrementQuality() {
-        item.quality++;
-    }
-
-    public void decrementQuality() {
-        item.quality--;
-    }
-
-    public void decrementQualityBy4() {
-        item.quality = max(item.quality - 4, 0);
-    }
-
-    public void decrementQualityBy2() {
-        item.quality = max(item.quality - 2, 0);
-    }
-
-    public void setLegendaryQuality() {
-        if (item.quality != LEGENDARY_ITEM_QUALITY) {
-            item.quality = LEGENDARY_ITEM_QUALITY;
-        }
-    }
-
     public boolean sellInLessThan5Days() {
         return item.sellIn >= 0 && item.sellIn <= 5;
     }
@@ -76,20 +53,32 @@ public class ItemHandler {
         return item.sellIn >= 5 && item.sellIn <= 10;
     }
 
-    public boolean sellInDaysMoreThan10Days() {
+    public boolean sellInMoreThan10Days() {
         return item.sellIn >= 10;
     }
 
-    public boolean qualityIsHigherThanZero() {
-        return item.quality > 0;
+    public void incrementQuality() {
+        item.quality++;
     }
 
-    public void makeQualityZero() {
-        item.quality = 0;
+    public void incrementQualityBy2() {
+        item.quality = max(item.quality + 2, 0);
     }
 
     public void incrementQualityBy3() {
         item.quality = max(item.quality + 3, 0);
+    }
+
+    public void decrementQuality() {
+        item.quality--;
+    }
+
+    public void decrementQualityBy2() {
+        item.quality = max(item.quality - 2, 0);
+    }
+
+    public void decrementQualityBy4() {
+        item.quality = max(item.quality - 4, 0);
     }
 
 }
