+++ Java/src/main/java/com/gildedrose/item_helpers/ItemType.java
@@ -1,7 +1,5 @@
 package com.gildedrose.item_helpers;
 
 public interface ItemType {
-
     void updateQuality();
-
 }
