+++ Java/src/main/java/com/gildedrose/items/LegendaryItem.java
@@ -1,11 +1,15 @@
 package com.gildedrose.items;
 
-import com.gildedrose.Item;
+import com.gildedrose.main.Item;
 import com.gildedrose.item_helpers.ItemHandler;
 import com.gildedrose.item_helpers.ItemType;
 
+import static com.gildedrose.item_helpers.ItemName.LEGENDARY;
+
 public class LegendaryItem implements ItemType {
 
+    public static final int LEGENDARY_ITEM_QUALITY = 80;
+    public static final String NOT_LEGENDARY_ITEM_ERROR_MESSAGE = "Item is legendary, quality must be always 80! Current value: ";
     private final ItemHandler item;
 
     public LegendaryItem(Item item) {
@@ -15,7 +19,10 @@ public class LegendaryItem implements ItemType {
     @Override
     public void updateQuality() {
         item.decrementSellInDate();
-        item.setLegendaryQuality();
+    }
+
+    public static boolean isNotLegendary(Item item) {
+        return item.name.equals(LEGENDARY.toString()) && item.quality != LEGENDARY_ITEM_QUALITY;
     }
 
 }
