+++ Java/src/main/java/com/gildedrose/main/Item.java
@@ -1,4 +1,4 @@
-package com.gildedrose;
+package com.gildedrose.main;
 
 public class Item {
 
@@ -14,8 +14,8 @@ public class Item {
         this.quality = quality;
     }
 
-   @Override
-   public String toString() {
+    @Override
+    public String toString() {
         return this.name + ", " + this.sellIn + ", " + this.quality;
     }
 }
